extract create fn
