+++ backend/es/conftest.py
@@ -6,9 +6,9 @@ import elasticsearch
 
 from addcorpus.python_corpora.load_corpus import load_corpus_definition
 from addcorpus.models import Corpus
-from es import es_index
 from es.models import Index
 from indexing.models import IndexJob, CreateIndexTask
+from indexing.run_create_task import create
 
 @pytest.fixture(scope='session')
 def mock_corpus():
@@ -78,7 +78,7 @@ def es_alias_client(db, es_server, es_client, mock_corpus, test_index_cleanup):
         index=index,
         delete_existing=True,
     )
-    es_index.create(create_task)
+    create(create_task)
     es_client.indices.create(index='test-times-2')
     es_client.indices.create(index='test-times-bla-3')
